requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.4.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isort==5.13.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
pexpect==4.8.0
platformdirs==4.2.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.6.4
pydantic_core==2.16.3
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==3.1.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.9
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.5.6
SQLAlchemy==2.0.28
ssh-import-id==5.11
starlette==0.36.3
systemd-python==234
tomli==2.0.1
tomlkit==0.12.4
Twisted==22.1.0
typing_extensions==4.10.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.28.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
zipp==1.0.0
zope.interface==5.4.0
